introduce `region` module
This modules is intended for all functionality that relates to
contiguous regions of guest memory. This differentiates it from
`guest_memory`, as that is about a holistic view of guest memory, and
from `mmap`, which is specifically about guest memory regions backed by
mmap VMAs.
